README.md
Put in descriptions for the files. Feel free to modify these or put them
under a new file name. I figured they'd be useful in some format, but
I'm not sure if they're best for a README file.  They'll need to be
edited later either way, but it's something useful for now.